Refactor to move all plan request actions and reducers to its own folders
