+++ client/src/store/actions/plans/requestNightWatchPlan.js
@@ -1,8 +1,8 @@
 import axios from 'axios';
 import addNightWatchPlanAction from './addNightWatchPlanAction';
-import fetchNightWatchPlanFailedAction from './party/fetchNightWatchPlanFailedAction';
-import switchOffLoadingPlanAction from './ui/switchOffLoadingPlanAction';
-import switchOnLoadingPlanAction from './ui/switchOnLoadingPlanAction';
+import fetchNightWatchPlanFailedAction from '../party/fetchNightWatchPlanFailedAction';
+import switchOffLoadingPlanAction from '../ui/switchOffLoadingPlanAction';
+import switchOnLoadingPlanAction from '../ui/switchOnLoadingPlanAction';
 
 const requestNightWatchPlan = ({ maxTotalTimeSpent, characters }) => (dispatch) => {
   dispatch(switchOnLoadingPlanAction());
