+++ client/src/store/reducer.js
@@ -1,7 +1,4 @@
-import * as actionTypes from './actions/actions';
 import { uiReducer } from './reducers/ui/reducer';
-import addNightWatchPlan from './reducers/addNightWatchPlan';
-import fetchNightWatchPlanFailed from './reducers/fetchNightWatchPlanFailed';
 import { DEFAULT_MAX_TIME_SPENT } from '../domain/NightWatchConfigEntity';
 import { MAX_TOTAL_TIME_SPENT_INCREASE } from './actions/config/actions';
 import { configReducer } from './reducers/config/reducer';
@@ -11,7 +8,9 @@ import {
   CHARACTER_REMOVE,
   CHARACTER_SLEEP_TIME_INCREASE,
 } from './actions/party/actions';
+import { ADD_NIGHT_WATCH_PLAN, FETCH_NIGHT_WATCH_PLAN_FAILED } from './actions/plans/actions';
 import { partyReducer } from './reducers/party/reducer';
+import { plansReducer } from './reducers/plans/reducer';
 
 const characters = [];
 export const initialState = () => ({
@@ -37,10 +36,9 @@ const reducer = (state = initialState(), action) => {
       return partyReducer(state, action);
     case MAX_TOTAL_TIME_SPENT_INCREASE:
       return configReducer(state, action);
-    case actionTypes.ADD_NIGHT_WATCH_PLAN:
-      return addNightWatchPlan(state, action);
-    case actionTypes.FETCH_NIGHT_WATCH_PLAN_FAILED:
-      return fetchNightWatchPlanFailed(state, action);
+    case ADD_NIGHT_WATCH_PLAN:
+    case FETCH_NIGHT_WATCH_PLAN_FAILED:
+      return plansReducer(state, action);
     default:
       return uiReducer(state, action);
   }
