+++ client/src/store/reducers/party/addCharacter.js
@@ -1,5 +1,5 @@
-import AddCharacterToParty from '../../usecases/party/AddCharacterToParty';
-import PartyRepository from '../../ports/party/PartyRepository';
+import AddCharacterToParty from '../../../usecases/party/AddCharacterToParty';
+import PartyRepository from '../../../ports/party/PartyRepository';
 
 const addCharacter = (state, { name }) => {
   const partyRepository = new PartyRepository(state);
